backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
supabase>=2,<3
python-dotenv>=1,<2
